created indexing package within cli folder
