+++ src/command_line/indexing/bulk_index.rs
@@ -1,5 +1,5 @@
 pub use super::prelude::*;
-use crate::command_line::split_path;
+use crate::command_line::indexing::split_path;
 use crate::indexing::index_manager::IndexManager;
 use crate::search::compound_processing::process_cpd;
 use bitvec::macros::internal::funty::Fundamental;
