+++ src/main.rs
@@ -20,19 +20,19 @@ async fn main() -> eyre::Result<()> {
     let app = Command::new("cheminee")
         .subcommand_required(true)
         .subcommand(command_line::fetch_pubchem::command())
-        .subcommand(command_line::index_sdf::command())
+        .subcommand(command_line::indexing::index_sdf::command())
         .subcommand(command_line::stream_pubchem_sdf::command())
         .subcommand(command_line::basic_search::command())
         .subcommand(rest_api::command())
         .subcommand(command_line::substructure_search::command())
-        .subcommand(command_line::create_index::command())
-        .subcommand(command_line::delete_index::command())
-        .subcommand(command_line::bulk_index::command());
+        .subcommand(command_line::indexing::create_index::command())
+        .subcommand(command_line::indexing::delete_index::command())
+        .subcommand(command_line::indexing::bulk_index::command());
 
     let matches = app.get_matches();
     let matches = match matches.subcommand().unwrap() {
-        (command_line::index_sdf::NAME, matches) => {
-            let writes = command_line::index_sdf::action(matches)?;
+        (command_line::indexing::index_sdf::NAME, matches) => {
+            let writes = command_line::indexing::index_sdf::action(matches)?;
             log::info!("wrote: {}", writes);
             Ok(())
         }
@@ -47,9 +47,15 @@ async fn main() -> eyre::Result<()> {
         (command_line::substructure_search::NAME, matches) => {
             command_line::substructure_search::action(matches)
         }
-        (command_line::create_index::NAME, matches) => command_line::create_index::action(matches),
-        (command_line::delete_index::NAME, matches) => command_line::delete_index::action(matches),
-        (command_line::bulk_index::NAME, matches) => command_line::bulk_index::action(matches),
+        (command_line::indexing::create_index::NAME, matches) => {
+            command_line::indexing::create_index::action(matches)
+        }
+        (command_line::indexing::delete_index::NAME, matches) => {
+            command_line::indexing::delete_index::action(matches)
+        }
+        (command_line::indexing::bulk_index::NAME, matches) => {
+            command_line::indexing::bulk_index::action(matches)
+        }
         (unknown, _) => panic!("🤨: {}", unknown),
     };
